习题至 4.3.10

+++ algs4/4.Graphs/4.3/Exercises/4.3.9.java
@@ -20,7 +20,7 @@ class EdgeWeightedGraph {
         for (int i = 0; i < E; i++) {
             int v = in.readInt();
             int w = in.readInt();
-            int weight = in.readInt();
+            double weight = in.readDouble();
             addEdge(new Edge(v, w, weight));
         }
     }
